moves search into a component
