+++ src/components/Home.js
@@ -17,7 +17,7 @@ const Home = () => {
         <img src={banner} className="banner w-25" alt="90s montage" />
       </section>
       <br></br>
-      <h4><em>New Releases Coming Soon:</em></h4>
+      <h4><em>New Releases Coming Soon!</em></h4>
       <img src={poster1} className="poster" alt="10 Things Movie Poster" />
       <img src={poster2} className="poster" alt="10 Things Movie Poster" />
       <img src={poster3} className="poster" alt="10 Things Movie Poster" />
